Extract SimpleUrlParser into separate Class
