+++ sdk/src/test/java/velocitas/sdk/parser/SimpleUrlParserTest.kt
@@ -14,19 +14,21 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package velocitas
+package velocitas.sdk.parser
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 
-class SimpleUrlParseTest : BehaviorSpec({
+class SimpleUrlParserTest : BehaviorSpec({
     context("Parsing NetLocation and Scheme") {
+        val simpleUrlParser = SimpleUrlParser()
+
         `when`("Parsing 'http://username:password@hostname:portnumber/path'") {
             val scheme = "http"
             val netLocation = "username:password@hostname:portnumber"
             val url = "$scheme://$netLocation/path"
 
-            val result = SimpleUrlParse(url)
+            val result = simpleUrlParser.parse(url)
 
             then("The scheme should be resolved to 'http'") {
                 result.scheme shouldBe scheme
@@ -41,7 +43,7 @@ class SimpleUrlParseTest : BehaviorSpec({
             val netLocation = "somehost:1234"
             val url = "$scheme://$netLocation/somePath"
 
-            val result = SimpleUrlParse(url)
+            val result = simpleUrlParser.parse(url)
 
             then("The scheme should be resolved to 'http'") {
                 result.scheme shouldBe scheme
@@ -56,7 +58,7 @@ class SimpleUrlParseTest : BehaviorSpec({
             val netLocation = "username:password@hostname:portnumber"
             val url = "$scheme://$netLocation/path"
 
-            val result = SimpleUrlParse(url)
+            val result = simpleUrlParser.parse(url)
 
             then("The scheme should be resolved in lowercase 'username:password@hostname:portnumber'") {
                 scheme.lowercase() shouldBe result.scheme
@@ -66,7 +68,7 @@ class SimpleUrlParseTest : BehaviorSpec({
         `when`("Parsing 'mailto:receipient@somewhere.io'") {
             val url = "mailto:receipient@somewhere.io"
 
-            val result = SimpleUrlParse(url)
+            val result = simpleUrlParser.parse(url)
 
             then("The scheme will be resolved to an empty String") {
                 result.scheme shouldBe ""
@@ -80,7 +82,7 @@ class SimpleUrlParseTest : BehaviorSpec({
             val netLocation = "127.0.0.1:42"
             val url = "//$netLocation/somePath"
 
-            val result = SimpleUrlParse(url)
+            val result = simpleUrlParser.parse(url)
 
             then("The scheme should be resolved to an empty String") {
                 result.scheme shouldBe ""
@@ -94,7 +96,7 @@ class SimpleUrlParseTest : BehaviorSpec({
             val netLocation = "localhost:123"
             val url = "$netLocation/somePath"
 
-            val result = SimpleUrlParse(url)
+            val result = simpleUrlParser.parse(url)
 
             then("The scheme should be resolved to an empty String") {
                 result.scheme shouldBe ""
